pet duration to minutes

+++ app/src/main/java/com/desireProj/ble_sdk/database/UploadETL.kt
@@ -8,7 +8,7 @@ data class UploadETL(
     @SerializedName("meetingDate")
     val day: String,
     @SerializedName("duration")
-    val duration: Long,
+    val duration: Float,
     @SerializedName("rssi")
     val rssi: Int,
     @SerializedName("uploadDate")
